# openapi.json
{
  "components": {
    "schemas": {
      "CreateNoteDto": {
        "properties": {
          "body": {
            "maxLength": 5000,
            "minLength": 1,
            "type": "string"
          },
          "status": {
            "default": "draft",
            "enum": [
              "draft",
              "published",
              "archived"
            ],
            "type": "string"
          },
          "title": {
            "maxLength": 200,
            "minLength": 1,
            "type": "string"
          }
        },
        "required": [
          "title",
          "body"
        ],
        "type": "object"
      },
      "NoteResponseDto": {
        "properties": {
          "_id": {
            "type": "string"
          },
          "body": {
            "type": "string"
          },
          "ownerId": {
            "type": "string"
          },
          "status": {
            "enum": [
              "draft",
              "published",
              "archived"
            ],
            "type": "string"
          },
          "title": {
            "type": "string"
          }
        },
        "required": [
          "_id",
          "title",
          "body",
          "status",
          "ownerId"
        ],
        "type": "object"
      },
      "UpdateNoteDto": {
        "properties": {
          "body": {
            "maxLength": 5000,
            "minLength": 1,
            "type": "string"
          },
          "status": {
            "default": "draft",
            "enum": [
              "draft",
              "published",
              "archived"
            ],
            "type": "string"
          },
          "title": {
            "maxLength": 200,
            "minLength": 1,
            "type": "string"
          }
        },
        "type": "object"
      }
    },
    "securitySchemes": {
      "accessToken": {
        "bearerFormat": "JWT",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "contact": {},
    "description": "",
    "title": "nestjs-template",
    "version": "0.1.0"
  },
  "openapi": "3.0.0",
  "paths": {
    "/api/healthz": {
      "get": {
        "operationId": "HealthController_check",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Health"
        ]
      }
    },
    "/api/v1/notes": {
      "get": {
        "operationId": "NotesController_list_v1",
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "maximum": 9007199254740991,
              "minimum": 1,
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "sortDirection",
            "required": false,
            "schema": {
              "default": "desc",
              "enum": [
                "asc",
                "desc"
              ],
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "enum": [
                "draft",
                "published",
                "archived"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A page of notes."
          }
        },
        "summary": "List notes",
        "tags": [
          "notes"
        ]
      },
      "post": {
        "operationId": "NotesController_create_v1",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateNoteDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NoteResponseDto"
                }
              }
            },
            "description": "The created note."
          }
        },
        "summary": "Create a note",
        "tags": [
          "notes"
        ]
      }
    },
    "/api/v1/notes/{id}": {
      "delete": {
        "operationId": "NotesController_remove_v1",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The note was deleted."
          },
          "404": {
            "description": "No note with that id exists."
          }
        },
        "summary": "Delete a note",
        "tags": [
          "notes"
        ]
      },
      "get": {
        "operationId": "NotesController_findById_v1",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NoteResponseDto"
                }
              }
            },
            "description": "The note."
          },
          "404": {
            "description": "No note with that id exists."
          }
        },
        "summary": "Get a note by id",
        "tags": [
          "notes"
        ]
      },
      "patch": {
        "operationId": "NotesController_update_v1",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateNoteDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NoteResponseDto"
                }
              }
            },
            "description": "The updated note."
          },
          "404": {
            "description": "No note with that id exists."
          }
        },
        "summary": "Update a note",
        "tags": [
          "notes"
        ]
      }
    }
  },
  "servers": [],
  "tags": []
}